+++ tests/test_particle.py
@@ -138,7 +138,7 @@ def test_underflow_u(one_p, seed):
 @pytest.mark.parametrize("seed", [0])
 def test_u_du_uwall_conversion(one_p, seed):
     one_p = random_p(one_p, seed)
-    ul, ur = sd.lagrangian._uleftright_from_udu(one_p.u, one_p.du, one_p.rx)
+    ul, ur = sd.utils._uleftright_from_udu(one_p.u, one_p.du, one_p.rx)
     u, du = u_du_from_uwall(ul, ur, one_p.rx)
     assert np.allclose(u, one_p.u)
     assert np.allclose(du, one_p.du)
